Move common deployment functionality into utils
