+++ tools/deploy.py
@@ -9,26 +9,9 @@ from eth_utils import to_checksum_address
 from raiden.log_config import configure_logging
 from raiden.network.rpc.client import JSONRPCClient
 from raiden.ui.cli import prompt_account
-from raiden.utils import get_contract_path
-from raiden.utils.solc import compile_files_cwd
+from raiden.utils.deployment import deploy_contracts
 
 log = structlog.get_logger(__name__)
-
-
-# Source files for all to be deployed solidity contracts
-RAIDEN_CONTRACT_FILES = [
-    'NettingChannelLibrary.sol',
-    'ChannelManagerLibrary.sol',
-    'Registry.sol',
-    'EndpointRegistry.sol',
-]
-
-# Top level contracts to be deployed. Dependencies are handled automatically
-# in `JSONRPCClient.deploy_solidity_contract()`
-CONTRACTS_TO_DEPLOY = [
-    'Registry.sol:Registry',
-    'EndpointRegistry.sol:EndpointRegistry',
-]
 
 
 def patch_deploy_solidity_contract():
@@ -82,35 +65,6 @@ def name_from_file(filename):
     return os.path.split(filename)[-1].partition('.')[0]
 
 
-def deploy_file(contract, compiled_contracts, client):
-    libraries = dict()
-    filename, _, name = contract.partition(":")
-    log.info(f"Deploying {name}")
-    proxy = client.deploy_solidity_contract(
-        name,
-        compiled_contracts,
-        libraries,
-        '',
-        contract_path=filename,
-    )
-
-    log.info(f"Deployed {name} @ {to_checksum_address(proxy.contract_address)}")
-    libraries[contract] = proxy.contract_address
-    return libraries
-
-
-def deploy_all(client):
-    contracts_expanded = [
-        get_contract_path(x)
-        for x in RAIDEN_CONTRACT_FILES
-    ]
-    compiled_contracts = compile_files_cwd(contracts_expanded)
-    deployed = {}
-    for contract in CONTRACTS_TO_DEPLOY:
-        deployed.update(deploy_file(contract, compiled_contracts, client))
-    return deployed
-
-
 def get_privatekey(keystore_path):
     address_hex, privatekey = prompt_account(None, keystore_path, None)
     return privatekey
@@ -138,14 +92,7 @@ def main(keystore_path, pretty, gas_price, port):
         gas_price_in_wei,
     )
 
-    deployed = deploy_all(client)
-
-    deployed['EndpointRegistry.sol:EndpointRegistry'] = to_checksum_address(
-        deployed['EndpointRegistry.sol:EndpointRegistry'],
-    )
-    deployed['Registry.sol:Registry'] = to_checksum_address(
-        deployed['Registry.sol:Registry'],
-    )
+    deployed = deploy_contracts(client)
 
     print(json.dumps(deployed, indent=2 if pretty else None))
